Add a new ReleaseNotesFetchers module
* ReleaseNotesRpmReader and ReleaseNotesUrlReader live now in that
  module as Rpm and Url classes.
