+++ test/lib/release_notes_fetchers/rpm_test.rb
@@ -1,10 +1,10 @@
 #!/usr/bin/env rspec
 
-require_relative "../test_helper"
-require "y2packager/release_notes_rpm_reader"
+require_relative "../../test_helper"
+require "y2packager/release_notes_fetchers/rpm"
 require "y2packager/product"
 
-describe Y2Packager::ReleaseNotesRpmReader do
+describe Y2Packager::ReleaseNotesFetchers::Rpm do
   subject(:reader) { described_class.new(product) }
 
   let(:product) { instance_double(Y2Packager::Product, name: "dummy") }
